api folytatás, errorLogger

+++ Etterem_backend/api/controllers/userController.js
@@ -3,30 +3,56 @@ const userRepository = require('../repositories/userRepository')
 exports.getUser = async (req,res,next) =>
 {
     try {
-        //const {index} = req.params
-        const index = 1
-        const user = await userRepository.getUser(index)
+        const id = 1
+        //const {id} = req.params
+        id = Number(id)
+        if(!id || isNaN(id)){
+            const error = new Error("Id not found or id is not a number!")
+            error.status = 404
+            throw error
+        }
+        const user = await userRepository.getUser(id)
         res.status(200).json(user)
     } catch (error) {
-        console.error("Error getting user:", error)
+        next(error)
     }
 }
 
-exports.getUsers = async (req,res,next) =>
+exports.getAllUser = async (req,res,next) =>
 {
     try {
         const users = await userRepository.getUsers()
         res.status(200).json(users)
     } catch (error) {
-        console.error("Error getting users:", error)
+        next(error)
     }
 }
 
 exports.createUser = async (req,res,next) =>
 {
     try {
+        const currentDate = new Date()
         //const {userName, fullName, email, password } = req.body
-        const currentDate = new Date();
+        // if(!userName){
+        //     const error = new Error("Username not found!")
+        //     error.status = 404
+        //     throw error
+        // }
+        // if(!fullName){
+        //     const error = new Error("Fullname not found!")
+        //     error.status = 404
+        //     throw error
+        // }
+        // if(!email){
+        //     const error = new Error("email not found!")
+        //     error.status = 404
+        //     throw error
+        // }
+        // if(!password){
+        //     const error = new Error("Password not found!")
+        //     error.status = 404
+        //     throw error
+        // }
         // const user = {
         //     id: null,
         //     timestamp: currentDate.toISOString(),
@@ -62,19 +88,27 @@ exports.createUser = async (req,res,next) =>
         }]
        
         await userRepository.createUser(users[0])
-        console.log("User created successfully")
-        res.status(201).send("User created successfully")
+        console.log("User created successfully!")
+        res.status(201).send("User created successfully!")
     } catch (error) {
-        console.error("Error creating user:", error)
+       next(error)
     }
 }
 
 exports.deleteUser = async (req,res,next) =>
 {
     try {
-        const {id} = req.params 
+        const {id} = req.params
+        id = Number(id)
+        if(!id || isNaN(id)){
+            const error = new Error("Id not found or id is not a number!")
+            error.status = 404
+            throw error
+        }
         await userRepository.deleteUser(id)
+        console.log("User deleted successfully!")
+        res.status(202).send("User deleted successfully!")
     } catch (error) {
-        console.error("error deleting user: ",error)
+        next(error)
     }
 }
